Fix #1z4na90 -- Password change not working correctly.
This fix actually fixes more than the misbehavior when changing
the password. Any change to a user that did not have
accountExpiry defined, would disable the account.

What happened is that if accountExpiry was not set, the program
would set it internally to "", and then use that value on its
next POST request to dhis2. But dhis2 interprets "" as setting
it to epoch (1970-01-01). Since accountExpiry gets thus set in
the past, the account appears as expired, and thus disabled.

This fix changes the model of the user so the field accountExpiry
becomes optional. This has repercussions in api and the codex too.