Fix #1z4na8y: org unit assignment not working.
In the user edit page, the selection of a org unit (in step 2) did
not work.

It turns out that the element OrgUnitSelectorFF its function onChange
expected the wrong kind of object. What OrgUnitSelector passes is a
list of strings with the path of the org units.

This fix takes a list of those paths and converts them into a list like
  [{ id: id1 }, { id: id2 }, ...]
which is what the rest of the code expects.