Attempt to add locales by extending the User model.
It is not trivial to add locale information (ui and db locales) to the
user, since that information is not stored in the same api endpoint. It
exists in "/userSettings", has no corresponding reading function in
d2-api, and has to be retrieved individually for each user (as opposed
to reading a batch of users in a single api call, which is possible).

This code attempts to add locale information by extending the model used
to represent a User. It is not working, but represents as far as I could
get by following this path.

The idea is to pass an extra parameter ("extraData") to the functions that
get and save information for a user ("getByIds()" and "save()"). If
extraData=true, there is an extra api call per user to read or write the
data that belong to userSettings (which are just keyUiLocale and keyDbLocale
for the moment).