Attempt to add locales by adding UserWithSettings.
It is not trivial to add locale information (ui and db locales) to the
user, since that information is not stored in the same api endpoint. It
exists in "/userSettings", has no corresponding reading function in
d2-api, and has to be retrieved individually for each user (as opposed
to reading a batch of users in a single api call, which is possible).

This code attempts to add locale information by creating a new model to
represent a single user, that we will be editing but not in bulk editing,
UserWithSettings. It is not working, but represents as far as I could
get by following this path.

The idea is to have the sigle-user edit form read and write from/to
UserWithSettings, instead of the old User[], and this one will already
know it has to do several api calls to read/write all its data
(ultimately coming from "/users" and "/userSettings").